// apps/web/lib/api/client/models/ReferralBatch.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ReferralBatch = {
    referral_batch_id: string;
    referral_batch_prefix: string;
    referral_batch_size: number;
    referral_outbound_facility_id: string;
    referral_inbound_facility_id: string;
};